requirements.txt
asgiref==3.4.1
autopep8==1.6.0
Django==4.0
django-crispy-forms==1.13.0
Pillow==8.4.0
pycodestyle==2.8.0
pycrypto==2.6.1
sqlparse==0.4.2
toml==0.10.2
